fix: restructure code into an MVC organization
