+++ app/src/main/java/com/example/allot/model/Entrant.java
@@ -1,4 +1,4 @@
-package com.example.allot;
+package com.example.allot.model;
 
 public class Entrant {
     private String deviceId;
