+++ app/src/main/java/com/example/allot/model/WaitingList.java
@@ -1,4 +1,4 @@
-package com.example.allot;
+package com.example.allot.model;
 
 import java.util.ArrayList;
 import java.util.HashMap;
