+++ app/src/main/java/com/example/allot/view/EventListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.allot;
+package com.example.allot.view;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,6 +8,8 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.allot.R;
 
 import java.util.ArrayList;
 import java.util.List;
